server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -6,9 +6,10 @@ const NotFoundError = require('../../exceptions/NotFoundError')
 const AuthorizationError = require('../../exceptions/AuthorizationError')
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool()
     this._collaborationService = collaborationService
+    this._cacheService = cacheService
   }
 
   async addNote({ title, body, tags, owner }) {
@@ -27,21 +28,37 @@ class NotesService {
       throw new InvariantError('Catatan gagal ditambahkan')
     }
 
+    // delete notes from cache after adding new note to database
+    await this._cacheService.delete(`notes:${owner}`)
+
     return result.rows[0].id
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-      LEFT JOIN collaborations ON collaborations.note_id = notes.id
-      WHERE notes.owner = $1 OR collaborations.user_id = $1
-      GROUP BY notes.id`,
-      values: [owner],
+    try {
+      // get notes from cache first
+      const result = await this._cacheService.get(`notes:${owner}`)
+      return JSON.parse(result)
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
+        LEFT JOIN collaborations ON collaborations.note_id = notes.id
+        WHERE notes.owner = $1 OR collaborations.user_id = $1
+        GROUP BY notes.id`,
+        values: [owner],
+      }
+
+      const result = await this._pool.query(query)
+      const mappedResult = result.rows.map(mapDBToModel)
+
+      // save notes to cache before return the result
+      await this._cacheService.set(
+        `notes:${owner}`,
+        JSON.stringify(mappedResult)
+      )
+
+      return mappedResult
     }
-
-    const result = await this._pool.query(query)
-
-    return result.rows.map(mapDBToModel)
   }
 
   async getNoteById(id) {
@@ -66,7 +83,7 @@ class NotesService {
     const updatedAt = new Date().toISOString()
 
     const query = {
-      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, created_at = $4 WHERE id = $5 RETURNING id',
+      text: 'UPDATE notes SET title = $1, body = $2, tags = $3, created_at = $4 WHERE id = $5 RETURNING id, owner',
       values: [title, body, tags, updatedAt, id],
     }
 
@@ -75,11 +92,15 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan')
     }
+
+    const { owner } = result.rows[0]
+    // delete notes from cache after editing notes from database
+    await this._cacheService.delete(`notes:${owner}`)
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
       values: [id],
     }
 
@@ -88,6 +109,10 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan')
     }
+
+    const { owner } = result.rows[0]
+    // delete notes from cache after editing notes from database
+    await this._cacheService.delete(`notes:${owner}`)
   }
 
   async verifyNoteOwner(id, owner) {
